добавлен: viewMogel, запрос яндекс погоды. минимально работоспособная версия

+++ app/src/main/java/com/example/weatherforecasts/yandexWeather/AnswerYandex.kt
@@ -6,18 +6,18 @@ data class AnswerYandex(
     val info: Info,
     val fact: Fact,
     val forecast: Forecast
-    )
+)
 
 data class Info(
     val lat: Float,
     val lon: Float,
     val url: String
-    )
+)
 
 data class Fact(
-    val temp: Float,
-    val feels_like: Float,
-    val temp_water: Float,
+    val temp: Int,
+    val feels_like: Int,
+    val temp_water: Int,
     val icon: String,
     val condition: String,
     val wind_speed: Float,
@@ -30,7 +30,7 @@ data class Fact(
     val polar: Boolean,
     val season: String,
     val obs_time: Int
-    )
+)
 
 data class Forecast(
     val date: String,
@@ -40,14 +40,14 @@ data class Forecast(
     val moon_code: Int,
     val moon_text: String,
     val parts: List<Parts>
-    )
+)
 
 data class Parts(
     val part_name: String,
-    val temp_min: Float,
-    val temp_max: Float,
-    val temp_avg: Float,
-    val feels_like: Float,
+    val temp_min: Int,
+    val temp_max: Int,
+    val temp_avg: Int,
+    val feels_like: Int,
     val icon: String,
     val condition: String,
     val daytime: String,
@@ -58,7 +58,7 @@ data class Parts(
     val pressure_mm: Float,
     val pressure_pa: Float,
     val humidity: Int,
-    val prec_mm: Float,
+    val prec_mm: Int,
     val prec_period: Int,
     val prec_prob: Float
-    )
+)
